requirements.txt
praw
pandas
nltk
streamlit
python-dotenv